+++ app/app_components/translateModule/services/translateConst_vmf.js
@@ -97,6 +97,7 @@ angular.module('translateModule')
                     none: 'alla wechhaun'
                 },
                 search: 'suchn'
-            }
+            },
+            desktopNotification: 'Desgdobb Benochrichtigunga'
         }
     });