+++ changelog.md
@@ -1,6 +1,7 @@
 
 ### Version 0.1.0-SNAPSHOT
 
+ - [feature] PRT 31 - enable/disable desktop notifications at user settings
  - [refactor] optimise favicon & add fry.png
  - [feature] replace bootstrap css with less
  - [refactor] optimise less mixins
@@ -90,4 +91,4 @@
  - [bug] Fixed Makefile
  - [initial] initial commit
 
--- Last change by: Tobias Arndt <tobias.arndt@devbliss.com> Wed Jun 03 16:37:06 CEST 2015
+-- Last change by: Marco Reinwarth <marco.reinwarth@devbliss.com> Wed Jun 03 11:49:27 CEST 2015